Move retracting to minting.rb
