+++ app/controllers/data_files_controller.rb
@@ -29,7 +29,6 @@ class DataFilesController < ApplicationController
   include Seek::Publishing::PublishingCommon
 
   include Seek::Doi::Minting
-  include Seek::Doi::Retracting
 
   include Seek::ISAGraphExtensions
