+++ app/controllers/documents_controller.rb
@@ -12,7 +12,6 @@ class DocumentsController < ApplicationController
   include Seek::Publishing::PublishingCommon
 
   include Seek::Doi::Minting
-  include Seek::Doi::Retracting
 
   include Seek::ISAGraphExtensions
