+++ app/controllers/presentations_controller.rb
@@ -12,7 +12,6 @@ class PresentationsController < ApplicationController
 
   include Seek::Publishing::PublishingCommon
   include Seek::Doi::Minting
-  include Seek::Doi::Retracting
 
   include Seek::ISAGraphExtensions
