+++ app/controllers/sops_controller.rb
@@ -13,7 +13,6 @@ class SopsController < ApplicationController
   include Seek::Publishing::PublishingCommon
 
   include Seek::Doi::Minting
-  include Seek::Doi::Retracting
 
   include Seek::ISAGraphExtensions
