+++ app/controllers/workflows_controller.rb
@@ -17,7 +17,6 @@ class WorkflowsController < ApplicationController
 
   include Seek::Publishing::PublishingCommon
   include Seek::Doi::Minting
-  include Seek::Doi::Retracting
   include Seek::ISAGraphExtensions
   include RoCrateHandling
   include Legacy::WorkflowSupport
